+++ Assets/Resources/Monster/orc_peon.prefab
@@ -727,9 +727,8 @@ GameObject:
   m_Component:
   - component: {fileID: 5870629044615271944}
   - component: {fileID: 5870629044624865704}
-  - component: {fileID: 6858376375774819484}
-  - component: {fileID: -2565424381512398344}
   - component: {fileID: 5627032518022052693}
+  - component: {fileID: 3488780952868328484}
   m_Layer: 0
   m_Name: orc_peon
   m_TagString: Monster
@@ -776,32 +775,6 @@ Animator:
   m_HasTransformHierarchy: 1
   m_AllowConstantClipSamplingOptimization: 1
   m_KeepAnimatorControllerStateOnDisable: 0
---- !u!114 &6858376375774819484
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5870629044615498280}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 32d8f77b42b42a246b5d2c8283e6ce76, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  monsterSpawner: {fileID: 0}
---- !u!114 &-2565424381512398344
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5870629044615498280}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bd362a9c7e11ef245aaae7b9bad59f3e, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  destination: []
 --- !u!195 &5627032518022052693
 NavMeshAgent:
   m_ObjectHideFlags: 0
@@ -824,6 +797,20 @@ NavMeshAgent:
   m_BaseOffset: 0
   m_WalkableMask: 4294967295
   m_ObstacleAvoidanceType: 4
+--- !u!114 &3488780952868328484
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5870629044615498280}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: bd362a9c7e11ef245aaae7b9bad59f3e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  monsterSpawner: {fileID: 0}
+  destination: []
 --- !u!1 &5870629044615498282
 GameObject:
   m_ObjectHideFlags: 0
