+++ Assets/Resources/MonsterDestination.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c96f062b24eb69c48bf45c057b9394b1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
