+++ Assets/Resources/MonsterDestination/Destination (1).prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9718191fdeb338148a7fe92ecce9115a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
