+++ Assets/Resources/MonsterDestination/Destination (2).prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c2c06e3388b7d8b48826eeb12009ec46
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
