+++ Assets/Resources/MonsterDestination/Destination (3).prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: dfc3ad028c1688b4caf9ca08eb95dbd8
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
