+++ Assets/Resources/MonsterDestination/Destination (4).prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4d65e6f2586b282439078e21e49717b7
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
