+++ Assets/Resources/MonsterDestination/Destination (5).prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ff072faea6d91e14db85a600f091a2b5
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
