+++ Assets/Resources/MonsterDestination/Destination.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8e2f515150598e143a140335c9d9ee71
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
