--- Assets/Scripts/Monster/MonsterDisappear.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 32d8f77b42b42a246b5d2c8283e6ce76
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
